requirements.txt
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_05nm_gqf36/croots/recipe/certifi_1663615689491/work/certifi
charset-normalizer==2.1.1
geographiclib==1.52
geopy==2.2.0
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
telebot==0.0.4
urllib3==1.26.12
